# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RLEGS AM Monitoring & Bot Telegram API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /am:
    get:
      operationId: listAccountManagers
      tags: [am]
      summary: List all account managers
      responses:
        "200":
          description: List of AMs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccountManager"
    post:
      operationId: createAccountManager
      tags: [am]
      summary: Create account manager
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountManagerBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountManager"
  /am/{id}:
    get:
      operationId: getAccountManager
      tags: [am]
      summary: Get account manager by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Account manager
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountManager"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateAccountManager
      tags: [am]
      summary: Update account manager
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountManagerBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountManager"
    delete:
      operationId: deleteAccountManager
      tags: [am]
      summary: Delete account manager
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /activity:
    get:
      operationId: listActivity
      tags: [activity]
      summary: Get activity data for all AMs
      parameters:
        - name: year
          in: query
          schema:
            type: ["integer", "null"]
        - name: month
          in: query
          schema:
            type: ["integer", "null"]
        - name: divisi
          in: query
          schema:
            type: ["string", "null"]
        - name: nama_am
          in: query
          schema:
            type: ["string", "null"]
        - name: label
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Activity data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityOverview"
  /activity/{nik}:
    get:
      operationId: getActivityByNik
      tags: [activity]
      summary: Get activity data for specific AM
      parameters:
        - name: nik
          in: path
          required: true
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: ["integer", "null"]
        - name: month
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Activity data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityDetail"
...
  /public/am/{slug}:
    get:
      operationId: getPublicAmProfile
      tags: [public]
      summary: Get public AM profile (requires NIK verification)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: nik
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: AM profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicAmProfile"
        "401":
          description: Invalid NIK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: AM not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        role:
          type: string
      required: [id, email, role]
    AccountManager:
      type: object
      properties:
        id:
          type: integer
        nik:
          type: string
        nama:
          type: string
        slug:
          type: string
        divisi:
          type: string
        segmen:
          type: ["string", "null"]
        witel:
          type: string
        telegramChatId:
          type: ["string", "null"]
        telegramConnected:
          type: boolean
        kpiActivity:
          type: integer
        createdAt:
          type: string
      required: [id, nik, nama, slug, divisi, witel, telegramConnected, kpiActivity, createdAt]
    CreateAccountManagerBody:
      type: object
      properties:
        nik:
          type: string
        nama:
          type: string
        divisi:
          type: string
        segmen:
          type: ["string", "null"]
        witel:
          type: string
        telegramChatId:
          type: ["string", "null"]
        kpiActivity:
          type: integer
      required: [nik, nama, divisi, witel]
    UpdateAccountManagerBody:
      type: object
      properties:
        nama:
          type: ["string", "null"]
        divisi:
          type: ["string", "null"]
        segmen:
          type: ["string", "null"]
        witel:
          type: ["string", "null"]
        telegramChatId:
          type: ["string", "null"]
        kpiActivity:
          type: ["integer", "null"]
...
    PerformanceSummary:
      type: object
      properties:
        nik:
          type: string
        namaAm:
          type: string
        divisi:
          type: string
        tahun:
          type: integer
        bulan:
          type: integer
        targetRevenue:
          type: number
        realRevenue:
          type: number
        achRate:
          type: number
        achRateYtd:
          type: number
        rankAch:
          type: integer
        statusWarna:
          type: string
        komponenDetail:
          type: ["string", "null"]
      required: [nik, namaAm, divisi, tahun, bulan, targetRevenue, realRevenue, achRate, achRateYtd, rankAch, statusWarna]
...
    FunnelLop:
      type: object
      properties:
        lopid:
          type: string
        judulProyek:
          type: string
        pelanggan:
          type: string
        nilaiProyek:
          type: number
        divisi:
          type: string
        statusF:
          type: string
        statusProyek:
          type: ["string", "null"]
        kategoriKontrak:
          type: ["string", "null"]
        estimateBulan:
          type: ["string", "null"]
        namaAm:
          type: string
        reportDate:
          type: string
      required: [lopid, judulProyek, pelanggan, nilaiProyek, divisi, statusF, namaAm, reportDate]
...
    FunnelDetail:
      type: object
      properties:
        nik:
          type: string
        namaAm:
          type: string
        divisi:
          type: string
        totalLop:
          type: integer
        totalNilai:
          type: number
        shortage:
          type: number
        lops:
          type: array
          items:
            $ref: "#/components/schemas/FunnelLop"
      required: [nik, namaAm, divisi, totalLop, totalNilai, shortage, lops]
    ActivityItem:
      type: object
      properties:
        id:
          type: integer
        nik:
          type: string
        fullname:
          type: string
        divisi:
          type: string
        activityEndDate:
          type: string
        label:
          type: string
        activityType:
          type: ["string", "null"]
        caName:
          type: ["string", "null"]
        lopid:
          type: ["string", "null"]
        picName:
          type: ["string", "null"]
        activityNotes:
          type: ["string", "null"]
      required: [id, nik, fullname, divisi, activityEndDate, label]
    ActivityOverview:
      type: object
      properties:
        totalActivity:
          type: integer
        byAm:
          type: array
          items:
            $ref: "#/components/schemas/ActivityAmGroup"
        activities:
          type: array
          items:
            $ref: "#/components/schemas/ActivityItem"
      required: [totalActivity, byAm, activities]
    ActivityAmGroup:
      type: object
      properties:
        nik:
          type: string
        fullname:
          type: string
        divisi:
          type: string
        activityCount:
          type: integer
        kpiTarget:
          type: integer
        kpiAchieved:
          type: boolean
        byLabel:
          type: array
          items:
            $ref: "#/components/schemas/ActivityLabelGroup"
      required: [nik, fullname, divisi, activityCount, kpiTarget, kpiAchieved, byLabel]
    ActivityLabelGroup:
      type: object
      properties:
        label:
          type: string
        count:
          type: integer
      required: [label, count]
    ActivityDetail:
      type: object
      properties:
        nik:
          type: string
        fullname:
          type: string
        divisi:
          type: string
        activityCount:
          type: integer
        kpiTarget:
          type: integer
        activities:
          type: array
          items:
            $ref: "#/components/schemas/ActivityItem"
      required: [nik, fullname, divisi, activityCount, kpiTarget, activities]
...
    PublicAmProfile:
      type: object
      properties:
        nik:
          type: string
        nama:
          type: string
        slug:
          type: string
        divisi:
          type: string
        witel:
          type: string
        performance:
          $ref: "#/components/schemas/PerformanceSummary"
        funnel:
          $ref: "#/components/schemas/FunnelDetail"
        activity:
          $ref: "#/components/schemas/ActivityDetail"
      required: [nik, nama, slug, divisi, witel]